feat(leave_request): add API for leave request

+++ src/main/java/com/lamdayne/humify/attendance/dto/request/UpdateLeaveBalanceRequest.java
@@ -4,6 +4,8 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PositiveOrZero;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class UpdateLeaveBalanceRequest {
     @NotNull(message = "LEAVE_TYPE_ID_REQUIRED")
@@ -14,5 +16,5 @@ public class UpdateLeaveBalanceRequest {
 
     @NotNull(message = "ALLOCATED_DAYS_REQUIRED")
     @PositiveOrZero(message = "ALLOCATED_DAYS_INVALID")
-    private Integer allocatedDays;
+    private BigDecimal allocatedDays;
 }
